Fixed menu button, file handling
Fixed the menu button going off into oblivion off-screen
Fixed the file picker not allowing to go outside of the app's files (code typo, whoops.)
Disabled v-sync by default - added as an option, just in case.
Added an FPS limiter - it was 60 fps by default before, but now you have the option to limit to 30 or even 15, in case you plan on running this on a smart fridge

Disabled the twitch tab to avoid confusion - it's not used for anything *yet*. In the future it will though!

The options introduced here are not persistent between sessions yet - it should be in the next update. I wanted to push this out for the button and file handler, the rest got caught up in it xD